requirements.txt
Django==4.2.10
psycopg2-binary==2.9.9
django-allauth==0.61.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
django-colorfield==0.11.0
pytest==7.4.3
pytest-django==4.6.0
pytest-cov==4.1.0
